% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lst_dds}
\alias{lst_dds}
\title{List of data dictionaries}
\format{
A list of data dictionaries for the NBDC Data Hub studies. First
level of the list is the study name (currently ABCD & HBCD), and
the second level is the version/release of the study.

\if{html}{\out{<div class="sourceCode">}}\preformatted{- lst_dds
    |- abcd
         |- 6.0
         |   |- data dictionary dataframe
         |- ...
    |- hbcd
         |- 1.0
         |   |- data dictionary dataframe
         |- ...
}\if{html}{\out{</div>}}
}
\usage{
lst_dds
}
\description{
List of data dictionaries
}
\keyword{datasets}
