% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lst_levels}
\alias{lst_levels}
\title{List of levels}
\format{
A list of categorical levels tables for the NBDC Data Hub studies.
First level of the list is the study name (currently ABCD & HBCD), and
the second level is the version/release of the study.

\if{html}{\out{<div class="sourceCode">}}\preformatted{- lst_levels
    |- abcd
         |- 6.0
         |   |- levels dataframe
         |- ...
    |- hbcd
         |- 1.0
         |   |- levels dataframe
         |- ...
}\if{html}{\out{</div>}}
}
\usage{
lst_levels
}
\description{
List of levels
}
\keyword{datasets}
