% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lst_session_latest}
\alias{lst_session_latest}
\title{List of latest session mappings}
\format{
A list mapping each study and release to the latest available
session metadata used by NBDCtools. First level of the list is the study
name (currently ABCD & HBCD), and the second level is the
version/release of the study. Each release contains:
\describe{
\item{table_name}{Character scalar with the table used for lookup.}
\item{session}{Character scalar with the latest session label.}
}

\if{html}{\out{<div class="sourceCode">}}\preformatted{- lst_session_latest
    |- abcd
         |- 6.0
         |   |- table_name: "ab_g_dyn"
         |   |- session: "ses-06A"
         |- ...
    |- hbcd
         |- 1.0
         |   |- table_name: "par_visit_data"
         |   |- session: "ses-V03"
         |- ...
}\if{html}{\out{</div>}}
}
\usage{
lst_session_latest
}
\description{
List of latest session mappings
}
\keyword{datasets}
