% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lst_sessions}
\alias{lst_sessions}
\title{List of sessions}
\format{
A list of session metadata for the NBDC Data Hub studies. First level
of the list is the study name (currently ABCD & HBCD), and the second level
is the version/release of the study.

\if{html}{\out{<div class="sourceCode">}}\preformatted{- lst_sessions
    |- abcd
         |- 6.0
         |   |- sessions dataframe
         |- ...
    |- hbcd
         |- 1.0
         |   |- sessions dataframe
         |- ...
}\if{html}{\out{</div>}}
}
\usage{
lst_sessions
}
\description{
List of sessions
}
\keyword{datasets}
