% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBDCtoolsData-package.R
\docType{package}
\name{NBDCtoolsData-package}
\alias{NBDCtoolsData}
\alias{NBDCtoolsData-package}
\title{NBDCtoolsData: NBDCtoolsData: NBDC Data Hub Tools (Data Package)}
\description{
This package provides data elements for the NBDCtools package. Data Hub.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://software.nbdc-datahub.org/NBDCtoolsData/}
}

}
\author{
\strong{Maintainer}: Janosch Linkersdoerfer \email{dairc.service@gmail.com} (\href{https://orcid.org/0000-0002-1577-1233}{ORCID})

Authors:
\itemize{
  \item Le Zhang \email{lezhang100@gmail.com} (\href{https://orcid.org/0009-0008-0205-2150}{ORCID})
}

}
\keyword{internal}
